RTSgBufInit: cSegs must be size_t so RT_ELEMENTS and similar will work without causing warnings.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@30470 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/iprt/sg.h
@@ -80,7 +80,7 @@ typedef PRTSGBUF *PPRTSGBUF;
  * @param   paSegs    Pointer to the start of the segment array.
  * @param   cSegs     Number of segments in the array.
  */
-RTDECL(void) RTSgBufInit(PRTSGBUF pSgBuf, PCRTSGSEG paSegs, unsigned cSegs);
+RTDECL(void) RTSgBufInit(PRTSGBUF pSgBuf, PCRTSGSEG paSegs, size_t cSegs);
 
 /**
  * Resets the internal buffer position of the S/G buffer to the beginning.
